pyproject.toml
[project]
name = "dynamicbot"
version = "0.1.0"
description = ""
authors = [
    {name = "Nikolai Trusov",email = "ntrusov@umschool.ru"}
]
readme = "README.md"
requires-python = ">=3.12,<4.0"
dependencies = [
    "pydantic (>=2.10.6,<3.0.0)",
    "dishka (>=1.4.2,<2.0.0)",
    "python-dotenv (>=1.0.1,<2.0.0)",
    "pydantic-settings (>=2.8.0,<3.0.0)",
    "python-json-logger (>=3.2.1,<4.0.0)",
    "aiogram-dialog (>=2.3.1,<3.0.0)",
    "fastapi[standard] (>=0.115.11,<0.116.0)",
    "faststream[rabbit] (>=0.5.35,<0.6.0)"

]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
